requirements.txt
beautifulsoup4==4.6.3
Pillow==5.3.0
pytesseract==0.2.5
